made the delete/report action thread relevant and permanent
